requirements.txt
requests
typing
datetime
pyyaml
dotenv
pillow
ephem
iCalendar